* kw/master: (24 commits)
  fixed typos
  added release notes
  added readme comments on how to run the examples
  added a readme about the examples
  camera map should use frame_id_t, not unsigned int
  static cast to type T to avoid msvc double to float cast warnings
  static cast to unsigned int to avoid msvc warnings
  use unique_ptr, not auto_ptr
  Remove boost install in Travis CI builds
  Catch all vital::io_exceptions and report error message
  prepend pos_dir in front of POS filenames
  added an example config for pos2krtd
  Removed the sample config files
  added example configs for maptk_analyze_tracks
  added example config files for public data sets
  update name of config file loaded by GUI
  default parameters for more algorithms
  update default number of iterations
  update CMake to glob all config files in the directory
  renamed/factored bundle_adjust configs
  ...

+++ maptk/plugins/ceres/bundle_adjust.cxx
@@ -483,8 +483,8 @@ bundle_adjust
         std::copy(d.begin(), d.begin()+num_dp, &intrinsic_params[5]);
       }
       // update the maps with the index of this new parameter vector
-      camera_intr_map[K] = camera_intr_params.size();
-      frame_to_intr_map[c.first] = camera_intr_params.size();
+      camera_intr_map[K] = static_cast<unsigned int>(camera_intr_params.size());
+      frame_to_intr_map[c.first] = static_cast<unsigned int>(camera_intr_params.size());
       // add the parameter vector
       camera_intr_params.push_back(intrinsic_params);
     }
